services/booking/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
asyncpg==0.30.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
pydantic==2.10.4
